Fix kernel entry point calculation
Currently for ABI v2 binaries we jump to the entry point in the
ELF header. For a vmlinux this works because the entry point is
0xc000000000000000 and the upper bits of the address are ignored
in real mode. For a zImage the entry point is 0x20000000 and as
a result kexec will jump to 0x20000000 + load_address which
typically results in dying in a fire. Fix this by turning the
entry point into an offset into the PT_LOAD section and jump
to that instead.

This patch allows us to kexec into a zImage directly, which
rules.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>