Use mem_top as the IMA size
The ePAPR entry ABI expects to be passed the amount of accessible memory
(aka the inital mapped area) in r7. The IMA concept only really makes sense
on BookE parts which always have the MMU enabled, but don't necessarily
have all of memory mapped. On PowerNV (where kexec-lite is mainly used)
we can always access all of memory in real mode.

The zImage will emit warnings if the dtb or initrd are outside the RMA so
we should pass a sane value for the IMA size from the trampoline. This
patch makes the trampoline pass the mem_top that was calculated when
building the kexec memory map as the IMA size.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>